Moves Options to Cli module
