+++ lib/tictactoe/cli/options/option.rb
